正解判定をAnswerResultに移動した #99
