+++ functions/test/utils.test.ts
@@ -3,30 +3,6 @@ import { assert } from "chai";
 
 import Utils from "../src/Utils";
 
-describe(".textSimilarity", () => {
-  const originalSentence =
-    "A New TV Show from The New York Times on FX and Hulu.";
-
-  it("比較されたテキストの類似度が返ること1", () => {
-    const targetSentence =
-      "A New TV Show from The New York Times on FX and Hulu";
-
-    assert.equal(Utils.textSimilarity(originalSentence, targetSentence), 100);
-  });
-
-  it("比較されたテキストの類似度が返ること2", () => {
-    const targetSentence = "A New TV Show from CNN on FX and Hulu";
-
-    assert.equal(Utils.textSimilarity(originalSentence, targetSentence), 79);
-  });
-
-  it("比較されたテキストの類似度が返ること3", () => {
-    const targetSentence = "A New TV Show from CNN on FX";
-
-    assert.equal(Utils.textSimilarity(originalSentence, targetSentence), 62);
-  });
-});
-
 describe(".findValueOfKeyInText", () => {
   const NEWS_SOURCES: { [key: string]: string } = {
     "cnn.com": "CNN",
